+++ src/test/java/com/oceanprotocol/squid/api/AssetsApiIT.java
@@ -1,12 +1,15 @@
 package com.oceanprotocol.squid.api;
 
 import com.fasterxml.jackson.core.type.TypeReference;
+import com.oceanprotocol.squid.api.config.OceanConfig;
+import com.oceanprotocol.squid.models.Account;
 import com.oceanprotocol.squid.models.DDO;
 import com.oceanprotocol.squid.models.DID;
 import com.oceanprotocol.squid.models.asset.AssetMetadata;
 import com.oceanprotocol.squid.models.asset.OrderResult;
 import com.oceanprotocol.squid.models.service.Service;
 import com.oceanprotocol.squid.models.service.ServiceEndpoints;
+import com.oceanprotocol.squid.core.sla.setup.SetupServiceAgreement;
 import com.typesafe.config.Config;
 import com.typesafe.config.ConfigFactory;
 import io.reactivex.Flowable;
@@ -20,6 +23,7 @@ import java.nio.file.Paths;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
+import java.util.Properties;
 
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertNotNull;
@@ -33,9 +37,9 @@ public class AssetsApiIT {
     private static String METADATA_JSON_CONTENT;
     private static AssetMetadata metadataBase;
     private static ServiceEndpoints serviceEndpoints;
-
-
     private static OceanAPI oceanAPI;
+    private static OceanAPI oceanAPIConsumer;
+
     private static  Config config;
 
     @BeforeClass
@@ -56,6 +60,29 @@ public class AssetsApiIT {
         assertNotNull(oceanAPI.getAssetsAPI());
         assertNotNull(oceanAPI.getMainAccount());
 
+        // TODO Enable the access's template registration in future versions
+        //setupServiceAgreement = new SetupServiceAgreement();
+        //setupServiceAgreement.registerTemplate();
+
+        Properties properties = new Properties();
+        properties.put(OceanConfig.KEEPER_URL, config.getString("keeper.url"));
+        properties.put(OceanConfig.KEEPER_GAS_LIMIT, config.getString("keeper.gasLimit"));
+        properties.put(OceanConfig.KEEPER_GAS_PRICE, config.getString("keeper.gasPrice"));
+        properties.put(OceanConfig.AQUARIUS_URL, config.getString("aquarius.url"));
+        properties.put(OceanConfig.SECRETSTORE_URL, config.getString("secretstore.url"));
+        properties.put(OceanConfig.CONSUME_BASE_PATH, config.getString("consume.basePath"));
+        properties.put(OceanConfig.MAIN_ACCOUNT_ADDRESS, config.getString("account.parity.address2"));
+        properties.put(OceanConfig.MAIN_ACCOUNT_PASSWORD,  config.getString("account.parity.password2"));
+        properties.put(OceanConfig.MAIN_ACCOUNT_CREDENTIALS_FILE, config.getString("account.parity.file2"));
+        properties.put(OceanConfig.DID_REGISTRY_ADDRESS, config.getString("contract.didRegistry.address"));
+        properties.put(OceanConfig.SERVICE_EXECUTION_AGREEMENT_ADDRESS, config.getString("contract.serviceExecutionAgreement.address"));
+        properties.put(OceanConfig.PAYMENT_CONDITIONS_ADDRESS,config.getString("contract.paymentConditions.address"));
+        properties.put(OceanConfig.ACCESS_CONDITIONS_ADDRESS, config.getString("contract.accessConditions.address"));
+        properties.put(OceanConfig.TOKEN_ADDRESS, config.getString("contract.token.address"));
+        properties.put(OceanConfig.DISPENSER_ADDRESS, config.getString("contract.dispenser.address"));
+
+        oceanAPIConsumer = OceanAPI.getInstance(properties);
+
     }
 
     @Test
@@ -70,15 +97,14 @@ public class AssetsApiIT {
 
     }
 
-
     @Test
     public void order() throws Exception {
 
         DDO ddo= oceanAPI.getAssetsAPI().create(metadataBase, serviceEndpoints);
         DID did= new DID(ddo.id);
-
-        Flowable<OrderResult> response = oceanAPI.getAssetsAPI().order(did, Service.DEFAULT_ACCESS_SERVICE_ID);
-
+      
+        Flowable<OrderResult> response = oceanAPIConsumer.getAssetsAPI().order(did, Service.DEFAULT_ACCESS_SERVICE_ID);
+      
         OrderResult result = response.blockingFirst();
         assertNotNull(result.getServiceAgreementId());
         assertEquals(true, result.isAccessGranted());
@@ -95,14 +121,14 @@ public class AssetsApiIT {
 
         log.debug("DDO registered!");
 
-        Flowable<OrderResult> response = oceanAPI.getAssetsAPI().order(did,  Service.DEFAULT_ACCESS_SERVICE_ID);
+        Flowable<OrderResult> response = oceanAPIConsumer.getAssetsAPI().order(did,  Service.DEFAULT_ACCESS_SERVICE_ID);
 
         OrderResult orderResult = response.blockingFirst();
         assertNotNull(orderResult.getServiceAgreementId());
         assertEquals(true, orderResult.isAccessGranted());
         log.debug("Granted Access Received for the service Agreement " + orderResult.getServiceAgreementId());
 
-        boolean result = oceanAPI.getAssetsAPI().consume(orderResult.getServiceAgreementId(), did, Service.DEFAULT_ACCESS_SERVICE_ID, "/tmp");
+        boolean result = oceanAPIConsumer.getAssetsAPI().consume(orderResult.getServiceAgreementId(), did, Service.DEFAULT_ACCESS_SERVICE_ID, "/tmp");
         assertEquals(true, result);
 
     }