+++ src/test/java/com/oceanprotocol/squid/api/SecretStoreApiIT.java
@@ -1,14 +1,19 @@
 package com.oceanprotocol.squid.api;
 
 import com.fasterxml.jackson.core.type.TypeReference;
+import com.oceanprotocol.keeper.contracts.AccessConditions;
+import com.oceanprotocol.squid.external.KeeperService;
+import com.oceanprotocol.squid.manager.ManagerHelper;
 import com.oceanprotocol.squid.models.DDO;
 import com.oceanprotocol.squid.models.DID;
 import com.oceanprotocol.squid.models.asset.AssetMetadata;
 import com.typesafe.config.Config;
 import com.typesafe.config.ConfigFactory;
+import com.typesafe.config.ConfigValueFactory;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 import org.junit.BeforeClass;
+import org.junit.Ignore;
 import org.junit.Test;
 
 import java.nio.file.Files;
@@ -25,31 +30,48 @@ public class SecretStoreApiIT {
     private static String METADATA_JSON_CONTENT;
     private static AssetMetadata metadataBase;
 
+    private static OceanAPI oceanAPIPublisher;
+    private static OceanAPI oceanAPIConsumer;
 
-    private static OceanAPI oceanAPI;
+    // This test only made sense when Secret Store `acl_contract` is set to "none"
+    // If not the complete encryption/decryption is tested in the purchase/order
 
     @BeforeClass
+    @Ignore
     public static void setUp() throws Exception {
 
         METADATA_JSON_CONTENT =  new String(Files.readAllBytes(Paths.get(METADATA_JSON_SAMPLE)));
         metadataBase = DDO.fromJSON(new TypeReference<AssetMetadata>() {}, METADATA_JSON_CONTENT);
 
         Config config = ConfigFactory.load();
-        oceanAPI = OceanAPI.getInstance(config);
+        oceanAPIPublisher = OceanAPI.getInstance(config);
 
-        assertNotNull(oceanAPI.getSecretStoreAPI());
-        assertNotNull(oceanAPI.getMainAccount());
+        String consumerAddress= config.getString("account.parity.address2");
+        String consumerPassword= config.getString("account.parity.password2");
+        String consumerFile= config.getString("account.parity.file2");
+
+        Config consumerConfig= config.withValue("account.main.address",
+                ConfigValueFactory.fromAnyRef(consumerAddress))
+                .withValue("account.main.password", ConfigValueFactory.fromAnyRef(consumerPassword))
+                .withValue("account.main.credentialsFile", ConfigValueFactory.fromAnyRef(consumerFile));
+
+        oceanAPIConsumer = OceanAPI.getInstance(consumerConfig);
+
+        assertNotNull(oceanAPIPublisher.getSecretStoreAPI());
+        assertNotNull(oceanAPIPublisher.getMainAccount());
 
     }
 
     @Test
+    @Ignore
     public void encrypt() throws Exception{
 
         String filesJson = metadataBase.toJson(metadataBase.base.files);
         String did = DID.builder().getHash();
 
-        String encryptedDocument = oceanAPI.getSecretStoreAPI().encrypt(did, filesJson, 0);
-        String decryptedDocument = oceanAPI.getSecretStoreAPI().decrypt(did, encryptedDocument);
+        String encryptedDocument = oceanAPIPublisher.getSecretStoreAPI().encrypt(did, filesJson, 0);
+
+        String decryptedDocument = oceanAPIConsumer.getSecretStoreAPI().decrypt(did, encryptedDocument);
 
         assertEquals(filesJson, decryptedDocument);