+++ CHANGELOG.md
@@ -1,6 +1,22 @@
 History
 =======
 
+0.3.3 (February 21th, 2019)
+-------------------------
+
+* Added Token Approve function
+* Using different Ocean's instances in the Integration Tests
+
+
+0.3.2 (February 14th, 2019)
+-------------------------
+
+* Added new features for DDO's Integrity
+* Convenience class to register Access Service's Template
+* Improvement to handle the events in order's flow
+* Support for Secret Store URLs coming from DDO
+* Fixed some bugs
+
 
 0.3.1 (February 8th, 2019)
 -------------------------